+++ e2e/pages/abstract_page.py
@@ -1,5 +1,8 @@
 from abc import ABC
+from dotenv import load_dotenv
 from playwright.sync_api import expect
+
+load_dotenv()
 
 
 class AbstractPage(ABC):