+++ e2e/uninstall.sh
@@ -1,0 +1,2 @@
+python3 -m playwright uninstall
+python3 -m pip uninstall -y -r requirements.txt